requirements.txt
torch
torchvision
pandas
nltk
rouge
numpy
tqdm
datasets
pyarrow
flask
flask-cors
pyyaml
